% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{methods}
\name{add_scenes}
\alias{add_scenes}
\alias{add_scenes,Blurb-method}
\title{Add scenes}
\usage{
add_scenes(object, scenes_list)

\S4method{add_scenes}{Blurb}(object, scenes_list)

\S4method{add_scenes}{Blurb}(object, scenes_list)
}
\description{
Add scenes
}
\seealso{
Other Package core functions: \code{\link{Blurb-class}},
  \code{\link{DataPipe-class}}, \code{\link{Scene-class}},
  \code{\link{Viz-class}}, \code{\link{generate_frames}},
  \code{\link{get_data_subset}}, \code{\link{get_data}},
  \code{\link{initialize,DataPipe-method}},
  \code{\link{make_blurb}}, \code{\link{prep_frames}},
  \code{\link{render_frames}}
}
\concept{Package core functions}
