% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{class}
\name{Blurb-class}
\alias{Blurb-class}
\alias{Blurb}
\title{Blurb class}
\description{
Blurb class
}
\seealso{
Other Package core functions: \code{\link{DataPipe-class}},
  \code{\link{Scene-class}}, \code{\link{Viz-class}},
  \code{\link{add_scenes}}, \code{\link{generate_frames}},
  \code{\link{get_data_subset}}, \code{\link{get_data}},
  \code{\link{initialize,DataPipe-method}},
  \code{\link{make_blurb}}, \code{\link{prep_frames}},
  \code{\link{render_frames}},
  \code{\link{set_template_fn}},
  \code{\link{show_data_schema}}
}
\concept{Package core functions}
