% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{class}
\name{DataPipe-class}
\alias{DataPipe-class}
\alias{DataPipe}
\title{DataPipe class}
\description{
DataPipe class
}
\seealso{
Other Package core functions: \code{\link{Scene-class}},
  \code{\link{Viz-class}}, \code{\link{get_data}},
  \code{\link{initialize,DataPipe-method}},
  \code{\link{prep_frames}}
}
\concept{Package core functions}
