% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{methods}
\name{generate_frames}
\alias{generate_frames}
\alias{generate_frames,Scene-method}
\title{Generate frames}
\usage{
generate_frames(object, chunking_type = "cumulative",
  chunking_fn = mean, ...)

\S4method{generate_frames}{Scene}(object, chunking_type = "cumulative",
  chunking_fn = mean, ...)
}
\description{
Generate frames
}
\seealso{
Other Package core functions: \code{\link{Blurb-class}},
  \code{\link{DataPipe-class}}, \code{\link{Scene-class}},
  \code{\link{Viz-class}}, \code{\link{add_scenes}},
  \code{\link{get_data_subset}}, \code{\link{get_data}},
  \code{\link{initialize,DataPipe-method}},
  \code{\link{make_blurb}}, \code{\link{prep_frames}},
  \code{\link{render_frames}}
}
\concept{Package core functions}
