% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{methods}
\name{get_data}
\alias{get_data}
\alias{get_data,DataPipe-method}
\title{Get Data}
\usage{
get_data(object)

\S4method{get_data}{DataPipe}(object)
}
\description{
Get Data
}
\seealso{
Other Package core functions: \code{\link{Blurb-class}},
  \code{\link{DataPipe-class}}, \code{\link{Scene-class}},
  \code{\link{Viz-class}}, \code{\link{add_scenes}},
  \code{\link{generate_frames}},
  \code{\link{get_data_subset}},
  \code{\link{initialize,DataPipe-method}},
  \code{\link{make_blurb}}, \code{\link{prep_frames}},
  \code{\link{render_frames}}
}
\concept{Package core functions}
