% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{methods}
\name{get_data_subset}
\alias{get_data_subset}
\alias{get_data_subset,DataPipe-method}
\title{Get Data Subset}
\usage{
get_data_subset(object, filter_fn, ...)

\S4method{get_data_subset}{DataPipe}(object,
  filter_fn = get_all_data_filter_fn, ...)
}
\description{
Get Data Subset
}
\seealso{
Other Package core functions: \code{\link{Blurb-class}},
  \code{\link{DataPipe-class}}, \code{\link{Scene-class}},
  \code{\link{Viz-class}}, \code{\link{add_scenes}},
  \code{\link{generate_frames}}, \code{\link{get_data}},
  \code{\link{initialize,DataPipe-method}},
  \code{\link{make_blurb}}, \code{\link{prep_frames}},
  \code{\link{render_frames}}
}
\concept{Package core functions}
