% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{methods}
\name{initialize,DataPipe-method}
\alias{initialize,DataPipe-method}
\alias{initialize,Scene-method}
\title{This is the constructor for the \link{DataPipe} class}
\usage{
\S4method{initialize}{DataPipe}(.Object, type = "in-memory",
  data_source_name)

\S4method{initialize}{Scene}(.Object, viz, data_pipe, num_frames,
  timeline_var)
}
\description{
DataPipe constructor

Scene constructor
}
\seealso{
Other Package core functions: \code{\link{DataPipe-class}},
  \code{\link{Scene-class}}, \code{\link{Viz-class}},
  \code{\link{get_data}}, \code{\link{prep_frames}}

Other Package core functions: \code{\link{DataPipe-class}},
  \code{\link{Scene-class}}, \code{\link{Viz-class}},
  \code{\link{get_data}}, \code{\link{prep_frames}}
}
\concept{Package core functions}
\keyword{internal}
