% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{methods}
\name{make_blurb}
\alias{make_blurb}
\alias{make_blurb,Blurb-method}
\title{Create a blurb}
\usage{
make_blurb(object, type = "gif", name = "blurb_output", ...)

\S4method{make_blurb}{Blurb}(object, type = "gif",
  name = "blurb_output", ...)
}
\description{
Create a blurb
}
\seealso{
Other Package core functions: \code{\link{Blurb-class}},
  \code{\link{DataPipe-class}}, \code{\link{Scene-class}},
  \code{\link{Viz-class}}, \code{\link{add_scenes}},
  \code{\link{generate_frames}},
  \code{\link{get_data_subset}}, \code{\link{get_data}},
  \code{\link{initialize,DataPipe-method}},
  \code{\link{prep_frames}}, \code{\link{render_frames}},
  \code{\link{set_template_fn}},
  \code{\link{show_data_schema}}
}
\concept{Package core functions}
