% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{methods}
\name{prep_frames}
\alias{prep_frames}
\alias{prep_frames,Scene-method}
\title{Prep frames}
\usage{
prep_frames(object)

\S4method{prep_frames}{Scene}(object)
}
\description{
Prep frames
}
\seealso{
Other Package core functions: \code{\link{DataPipe-class}},
  \code{\link{Scene-class}}, \code{\link{Viz-class}},
  \code{\link{get_data}},
  \code{\link{initialize,DataPipe-method}}
}
\concept{Package core functions}
