% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{methods}
\name{render_frames}
\alias{render_frames}
\alias{render_frames,Scene-method}
\title{Render frames}
\usage{
render_frames(object, ...)

\S4method{render_frames}{Scene}(object, ...)
}
\description{
Render frames
}
\seealso{
Other Package core functions: \code{\link{Blurb-class}},
  \code{\link{DataPipe-class}}, \code{\link{Scene-class}},
  \code{\link{Viz-class}}, \code{\link{add_scenes}},
  \code{\link{generate_frames}},
  \code{\link{get_data_subset}}, \code{\link{get_data}},
  \code{\link{initialize,DataPipe-method}},
  \code{\link{make_blurb}}, \code{\link{prep_frames}}
}
\concept{Package core functions}
