% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{methods}
\name{set_template_fn}
\alias{set_template_fn}
\alias{set_template_fn,DataPipe-method}
\title{Set template function}
\usage{
set_template_fn(object, template_fn)

\S4method{set_template_fn}{DataPipe}(object, template_fn)
}
\description{
Set template function
}
\seealso{
Other Package core functions: \code{\link{Blurb-class}},
  \code{\link{DataPipe-class}}, \code{\link{Scene-class}},
  \code{\link{Viz-class}}, \code{\link{add_scenes}},
  \code{\link{generate_frames}},
  \code{\link{get_data_subset}}, \code{\link{get_data}},
  \code{\link{initialize,DataPipe-method}},
  \code{\link{make_blurb}}, \code{\link{prep_frames}},
  \code{\link{render_frames}},
  \code{\link{show_data_schema}}
}
\concept{Package core functions}
