% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{methods}
\name{show_data_schema}
\alias{show_data_schema}
\alias{show_data_schema,DataPipe-method}
\title{Show data schema}
\usage{
show_data_schema(object)

\S4method{show_data_schema}{DataPipe}(object)
}
\description{
Show data schema
}
\seealso{
Other Package core functions: \code{\link{Blurb-class}},
  \code{\link{DataPipe-class}}, \code{\link{Scene-class}},
  \code{\link{Viz-class}}, \code{\link{add_scenes}},
  \code{\link{generate_frames}},
  \code{\link{get_data_subset}}, \code{\link{get_data}},
  \code{\link{initialize,DataPipe-method}},
  \code{\link{make_blurb}}, \code{\link{prep_frames}},
  \code{\link{render_frames}},
  \code{\link{set_template_fn}}
}
\concept{Package core functions}
