% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{class}
\name{Viz-class}
\alias{Viz-class}
\alias{Viz}
\title{Viz class}
\description{
Viz class
}
\seealso{
Other Package core functions: \code{\link{DataPipe-class}},
  \code{\link{Scene-class}}, \code{\link{get_data}},
  \code{\link{initialize,DataPipe-method}},
  \code{\link{prep_frames}}
}
\concept{Package core functions}
