% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{class}
\name{Scene-class}
\alias{Scene-class}
\alias{Scene}
\title{Scene class}
\description{
Scene class
}
\seealso{
Other Package core functions: \code{\link{Blurb-class}},
  \code{\link{DataPipe-class}}, \code{\link{Viz-class}},
  \code{\link{add_scenes}}, \code{\link{generate_frames}},
  \code{\link{get_data_subset}}, \code{\link{get_data}},
  \code{\link{initialize,DataPipe-method}},
  \code{\link{make_blurb}}, \code{\link{prep_frames}},
  \code{\link{render_frames}}
}
\concept{Package core functions}
